+++ src/main/java/com/mindbloom/dao/MentalHealthResourceDaoHibernate.java
@@ -33,7 +33,7 @@ public class MentalHealthResourceDaoHibernate implements MentalHealthResourceDao
        DELETE
        =========================== */
     @Override
-    public void delete(Long id) {
+    public void delete(int id) {
         MentalHealthResource resource = findById(id);
         if (resource != null) {
             getSession().delete(resource);
@@ -44,7 +44,7 @@ public class MentalHealthResourceDaoHibernate implements MentalHealthResourceDao
        FIND BY ID
        =========================== */
     @Override
-    public MentalHealthResource findById(Long id) {
+    public MentalHealthResource findById(int id) {
         return getSession().get(MentalHealthResource.class, id);
     }
